requirements.txt
Flask==3.0.3
PyMySQL==1.1.1
selenium==4.8.2
requests==2.28.1
urllib3==1.26.13

